+++ cello.h
@@ -35,8 +35,10 @@ BEGIN_JUCE_MODULE_DECLARATION
 END_JUCE_MODULE_DECLARATION
 */
 
+#include "cello/cello_ipc.h"
 #include "cello/cello_object.h"
 #include "cello/cello_path.h"
 #include "cello/cello_query.h"
+#include "cello/cello_sync.h"
 #include "cello/cello_update_source.h"
 #include "cello/cello_value.h"